+++ app/src/main/java/com/fpf/smartscan/ui/components/modals/SelectorModal.kt
@@ -1,4 +1,4 @@
-package com.fpf.smartscan.ui.components
+package com.fpf.smartscan.ui.components.modals
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -11,6 +11,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
+import com.fpf.smartscan.ui.components.SelectorItem
 import kotlinx.coroutines.FlowPreview
